Fix ivec

+++ src/core/vertex.hpp
@@ -14,9 +14,9 @@ namespace slt
     using vec3 = glm::vec3;
     using vec2 = glm::vec2;
 
-    using ivec4 = glm::vec4;
-    using ivec3 = glm::vec3;
-    using ivec2 = glm::vec2;
+    using ivec4 = glm::ivec4;
+    using ivec3 = glm::ivec3;
+    using ivec2 = glm::ivec2;
 
     using mat4  = glm::mat4;
     using mat3  = glm::mat3;
